Reordered steps in Scrum 2 and Scrum 3
Prioritized tasks based on Astro Teller's advice to focus first on the problems which present the most learning challenges. 
Added to Scrum 2: IDE updating objects in active library (previously Scrum 3)
Added to Scrum 3: Global object inclusion interface (previously Scrum 2)
Updating an objects code in game is at the core of the project. Managing the object libraries included in the game is peripheral UI frills for later.